requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-birthday==0.1.3
django-crispy-forms==1.8.1
gunicorn==20.0.4
idna==2.8
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
